requirements.txt
streamlit==1.35.0
google-generativeai==0.3.2
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==0.0.6
PyPDF2==3.0.1
faiss-cpu>=1.9.0
numpy
tqdm
pandas

